pyproject.toml
[tool.poetry]
name = "yaml-source-map"
version = "0.1.0"
description = "Calculate JSON Pointers to each value within a YAML document along with the line, column and character position for the start and end of that value"
authors = ["David Andersson <anderssonpublic@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
black = "^21.5b1"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
